+++ extensions/iceworks-app/src/commands/executeCommand.ts
@@ -1,28 +1,24 @@
 import * as vscode from 'vscode';
 import { Terminal, TerminalOptions } from 'vscode';
-import { ITerminalMap } from '../types';
 
-export default function executeCommand(terminalMapping: ITerminalMap, script: vscode.Command, id?: string) {
-  if (!script.arguments) {
+export default function executeCommand(command: vscode.Command) {
+  if (!command.arguments) {
     return;
   }
-  const args = script.arguments;
-  const [cwd, command] = args;
-  if (!command) {
+  const args = command.arguments;
+  const [cwd, script] = args;
+  if (!script) {
     return;
   }
 
-  const terminalId = id || command;
-  let terminal: Terminal;
+  const terminals = vscode.window.terminals;
 
-  if (terminalMapping.has(terminalId)) {
-    terminal = terminalMapping.get(terminalId)!;
-  } else {
-    const terminalOptions: TerminalOptions = { cwd, name: command };
+  let terminal: Terminal | undefined = terminals.find((terminal: Terminal) => terminal.name === script);
+  if (!terminal) {
+    const terminalOptions: TerminalOptions = { cwd, name: script };
     terminal = vscode.window.createTerminal(terminalOptions);
-    terminalMapping.set(terminalId, terminal);
   }
 
   terminal.show();
-  terminal.sendText(command);
+  terminal.sendText(script);
 }